I fixed the conditional expression that needs to calculate the difference using the second argument of the `difference` function.
A situation where the difference between the cumulative sum of staking reward amount cannot be used for calculation is when all of the following conditions are true:

- The cumulative staking amount on the protocol total equals the cumulative staking amount on the Property (= 100% of historical staking share)
- The Property staking amount equals the user staking amount

When these are satisfied, 100% of the cumulative sum of the maximum mint amount is allocated to the relevant Property, and 100% thereof is allocated to one user. Originally, the user should only get increments from the staking block, but this user receives all the rewards. To avoid this problem, it uses the second argument of the `difference` function to calculate the difference.